requirements.txt
fastapi
uvicorn[standard]
openai
requests
python-dotenv
xdk
pyttsx3
SpeechRecognition
yt-dlp
